dashboard

+++ src/pages/ManageService/ManageService.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react';
-
+import { Link } from 'react-router-dom';
 
 const ManageService = () => {
     const [services, setServices] = useState([])
@@ -9,42 +9,58 @@ const ManageService = () => {
             .then(data => setServices(data))
     }, []);
 
-    const handleFelete = id => {
-        const url = `http://localhost:5000/orders/${id}`;
-        fetch(url, {
-            method: 'DELETE'
-        })
-            .then(res => res.json())
-            .then(data => {
-                console.log(data)
-                if (data.deletedCount) {
-                    alert('deleted')
-                    const remaining = services.filter(service => service._id !== id)
-                    setServices(remaining)
-                }
-
+    const handleDeleteUser = id => {
+        const proceed = window.confirm('Are you sure, you want to delete?');
+        if (proceed) {
+            const url = `http://localhost:5000/orders/${id}`;
+            fetch(url, {
+                method: 'DELETE'
             })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount > 0) {
+                        alert('deleted successfully');
+                        const remainingUsers = services.filter(service => service._id !== id);
+                        setServices(remainingUsers);
+                    }
+                });
+        }
     }
     return (
+        // <div>
+        //     <h4>manage services</h4>
+        //     <li>
+
+        //         {
+        //             services.map(services => <div
+        //                 key={services.id}
+
+
+        //             >
+        //                 <h4>{services.name}</h4>
+        //                 <button onClick={() => handleFelete(services._id)}>Delete</button>
+
+        //             </div>)
+        //         }
+        //     </li>
+
+        // </div>
         <div>
-            <h4>manage services</h4>
-            <li>
-
+            <h2>Users Available: {services.length} </h2>
+            <ul>
                 {
-                    services.map(services => <div
-                        key={services.id}
-
-
-                    >
-                        <h4>{services.name}</h4>
-                        <button onClick={() => handleFelete(services._id)}>Delete</button>
-
-                    </div>)
+                    services.map(services => <li
+                        key={services._id}
+                    >{services.productName} : {services.description}
+                        <Link to={`/users/update/${services._id}`}><button>Update</button></Link>
+                        <button onClick={() => handleDeleteUser(services._id)}>X</button>
+                    </li>)
                 }
-            </li>
-
+            </ul>
         </div>
     );
 };
+
+
 
 export default ManageService;
